New updates for deep NMF with SS constraints
This revision allows to use new updates for deep NMF with SS constraints inspired by the works of C.Fevotte and D. Dobigeon.
The notebook "notebook_deep_nmf_structure" shows how to call the two existing updates, that are the one based on Lagrangian multipliers (with NR or Bisec routines) or the ones detailed above.